BUG: error in create_spatial_index on a read-only file even if index exists already (#686)
If the index already exists, it is more practical that
`create_spatial_index` doesn't fail if the file is not writable so it is
not needed to first explicitly check if the index exists.

This was also the behaviour of geofileops < 0.10.

+++ geofileops/fileops.py
@@ -613,12 +613,15 @@ def create_spatial_index(
     if exist_ok and force_rebuild:
         raise ValueError("exist_ok and force_rebuild can't both be True")
 
-    # .gpkg.zip files don't support update, so use has_spatial_index up-front to check
-    # if there is an index.
-    if str(path).lower().endswith(".gpkg.zip"):
+    # Add index
+    path_info = _geofileinfo.get_geofileinfo(path)
+    try:
+        # Use has_spatial_index up-front to check if there is an index. To avoid needing
+        # R/W permissions, don't open the file in update mode yet.
+        remove_spatial_index_needed = False
         if has_spatial_index(path, layer):
             if force_rebuild:
-                pass
+                remove_spatial_index_needed = True
             elif exist_ok:
                 return
             else:
@@ -626,26 +629,12 @@ def create_spatial_index(
                     f"spatial index already exists on {path}#{layer.name}"
                 )
 
-        raise RuntimeError(
-            f"create_spatial_index not supported for .gpkg.zip files: {path}"
-        )
-
-    # Add index
-    path_info = _geofileinfo.get_geofileinfo(path)
-    try:
         # The config options need to be set before opening the file!
         with _ogr_util.set_config_options({"OGR_SQLITE_CACHE": cache_size_mb}):
             datasource = gdal.OpenEx(str(path), nOpenFlags=gdal.OF_UPDATE)
             # If index already exists, remove index or return
-            if has_spatial_index(path, layer, datasource=datasource):
-                if force_rebuild:
-                    remove_spatial_index(path, layer, datasource=datasource)
-                elif exist_ok:
-                    return
-                else:
-                    raise RuntimeError(
-                        f"spatial index already exists on {path}#{layer.name}"
-                    )
+            if remove_spatial_index_needed:
+                remove_spatial_index(path, layer, datasource=datasource)
 
             if path_info.is_spatialite_based:
                 geometrycolumn = layer.geometrycolumn
@@ -1451,81 +1440,103 @@ def _read_file_base_fiona(
         layer = get_only_layer(path)
 
     # VERY DIRTY hack to get the fid
-    if fid_as_index:
-        # Make a copy/copy input file to geopackage, as we will add an fid/rowd column
-        tmp_fid_path = Path(tempfile.mkdtemp()) / f"{Path(path).stem}.gpkg"
-        path_info = _geofileinfo.get_geofileinfo(path)
-        try:
+
+    tmp_fid_path = None
+    try:
+        if fid_as_index:
+            # Make a copy/copy input file to geopackage,
+            # as we will add an fid/rowd column
+            tmp_fid_path = Path(tempfile.mkdtemp()) / f"{Path(path).stem}.gpkg"
+            path_info = _geofileinfo.get_geofileinfo(path)
             if path_info.driver == "GPKG":
-                copy(path, tmp_fid_path)
+                copy(path, tmp_fid_path, keep_permissions=False)
             else:
                 copy_layer(path, tmp_fid_path)
+            fid_column = get_layerinfo(
+                tmp_fid_path, layer=layer, raise_on_nogeom=False
+            ).fid_column
             if path_info.is_fid_zerobased:
                 # fid in shapefile is 0 based, so fid-1
-                add_column(tmp_fid_path, "__TMP_GEOFILEOPS_FID", "INTEGER", "fid-1")
+                add_column(
+                    tmp_fid_path,
+                    "__TMP_GEOFILEOPS_FID",
+                    "INTEGER",
+                    f"{fid_column}-1",
+                    layer=layer,
+                )
             else:
-                add_column(tmp_fid_path, "__TMP_GEOFILEOPS_FID", "INTEGER", "fid")
+                add_column(
+                    tmp_fid_path,
+                    "__TMP_GEOFILEOPS_FID",
+                    "INTEGER",
+                    fid_column,
+                    layer=layer,
+                )
 
             path = tmp_fid_path
-        finally:
-            if ConfigOptions.remove_temp_files and tmp_fid_path.parent.exists():
-                shutil.rmtree(tmp_fid_path, ignore_errors=True)
 
-    # Checking if field/column names should be read is case sensitive in fiona, so
-    # make sure the column names specified have the same casing.
-    columns_prepared = None
-    if columns is not None:
-        layerinfo = get_layerinfo(path, layer=layer, raise_on_nogeom=False)
-        columns_upper_lookup = {column.upper(): column for column in columns}
-        columns_prepared = {
-            column: columns_upper_lookup[column.upper()]
-            for column in layerinfo.columns
-            if column.upper() in columns_upper_lookup
-        }
+        # Checking if field/column names should be read is case sensitive in fiona, so
+        # make sure the column names specified have the same casing.
+        columns_prepared = None
+        if columns is not None:
+            layerinfo = get_layerinfo(path, layer=layer, raise_on_nogeom=False)
+            columns_upper_lookup = {column.upper(): column for column in columns}
+            columns_prepared = {
+                column: columns_upper_lookup[column.upper()]
+                for column in layerinfo.columns
+                if column.upper() in columns_upper_lookup
+            }
 
-    # Read...
-    columns_list = None if columns_prepared is None else list(columns_prepared)
-    result_gdf = gpd.read_file(
-        str(path),
-        layer=layer,
-        bbox=bbox,
-        rows=rows,
-        columns=columns_list,
-        where=where,
-        sql=sql_stmt,
-        sql_dialect=sql_dialect,
-        ignore_geometry=ignore_geometry,
-        **kwargs,
-    )
+        # Read...
+        columns_list = None if columns_prepared is None else list(columns_prepared)
+        result_gdf = gpd.read_file(
+            str(path),
+            layer=layer,
+            bbox=bbox,
+            rows=rows,
+            columns=columns_list,
+            where=where,
+            sql=sql_stmt,
+            sql_dialect=sql_dialect,
+            ignore_geometry=ignore_geometry,
+            **kwargs,
+        )
 
-    # Set the index to the backed-up fid
-    if fid_as_index:
-        result_gdf = result_gdf.set_index("__TMP_GEOFILEOPS_FID")
-        result_gdf.index.name = "fid"
+        # Set the index to the backed-up fid
+        if fid_as_index:
+            result_gdf = result_gdf.set_index("__TMP_GEOFILEOPS_FID")
+            result_gdf.index.name = "fid"
 
-    # Reorder columns + change casing so they are the same as columns parameter
-    if columns_prepared is not None and len(columns_prepared) > 0:
-        columns_to_keep = list(columns_prepared)
-        if "geometry" in result_gdf.columns:
-            columns_to_keep += ["geometry"]
-        result_gdf = result_gdf[columns_to_keep]
-        result_gdf = result_gdf.rename(columns=columns_prepared)
+        # Reorder columns + change casing so they are the same as columns parameter
+        if columns_prepared is not None and len(columns_prepared) > 0:
+            columns_to_keep = list(columns_prepared)
+            if "geometry" in result_gdf.columns:
+                columns_to_keep += ["geometry"]
+            result_gdf = result_gdf[columns_to_keep]
+            result_gdf = result_gdf.rename(columns=columns_prepared)
 
-    # Starting from fiona 1.9, string columns with all None values are read as being
-    # float columns. Convert them to object type.
-    float_cols = list(result_gdf.select_dtypes(["float64"]).columns)
-    if len(float_cols) > 0:
-        # Check for all float columns found if they should be object columns instead
-        import fiona
+        # Starting from fiona 1.9, string columns with all None values are read as being
+        # float columns. Convert them to object type.
+        float_cols = list(result_gdf.select_dtypes(["float64"]).columns)
+        if len(float_cols) > 0:
+            # Check for all float columns found if they should be object columns instead
+            import fiona
 
-        with fiona.open(path, layer=layer) as collection:
-            assert collection.schema is not None
-            properties = collection.schema["properties"]
-            for col in float_cols:
-                if col in properties and properties[col].startswith("str"):
-                    result_gdf[col] = (
-                        result_gdf[col].astype(object).replace(np.nan, None)
-                    )
+            with fiona.open(path, layer=layer) as collection:
+                assert collection.schema is not None
+                properties = collection.schema["properties"]
+                for col in float_cols:
+                    if col in properties and properties[col].startswith("str"):
+                        result_gdf[col] = (
+                            result_gdf[col].astype(object).replace(np.nan, None)
+                        )
+    finally:
+        if (
+            tmp_fid_path is not None
+            and ConfigOptions.remove_temp_files
+            and tmp_fid_path.parent.exists()
+        ):
+            shutil.rmtree(tmp_fid_path.parent, ignore_errors=True)
 
     return result_gdf
 
@@ -2220,7 +2231,11 @@ def cmp(
         return filecmp.cmp(str(path1), str(path2))
 
 
-def copy(src: Union[str, "os.PathLike[Any]"], dst: Union[str, "os.PathLike[Any]"]):
+def copy(
+    src: Union[str, "os.PathLike[Any]"],
+    dst: Union[str, "os.PathLike[Any]"],
+    keep_permissions: bool = True,
+):
     """Copies the geofile from src to dst.
 
     If the source file is a geofile containing of multiple files (eg. .shp) all files
@@ -2229,6 +2244,8 @@ def copy(src: Union[str, "os.PathLike[Any]"], dst: Union[str, "os.PathLike[Any]"
     Args:
         src (PathLike): the file to copy. |GDAL_vsi| paths are not supported.
         dst (PathLike): the location to copy the file(s) to.
+        keep_permissions (bool, optional): True to keep the file permissions of the
+            source file. Defaults to True.
 
     .. |GDAL_vsi| raw:: html
 
@@ -2244,21 +2261,27 @@ def copy(src: Union[str, "os.PathLike[Any]"], dst: Union[str, "os.PathLike[Any]"
     src_info = _geofileinfo.get_geofileinfo(src)
 
     # Copy the main file
-    shutil.copy(str(src), dst)
+    if keep_permissions:
+        shutil.copy(str(src), dst)
+    else:
+        dstfile = dst / src.name if dst.is_dir() else dst
+        shutil.copyfile(src, dstfile)
 
     # For some file types, extra files need to be copied
     # If dest is a dir, just use move. Otherwise concat dest filepaths
-    if dst.is_dir():
-        for suffix in src_info.suffixes_extrafiles:
-            srcfile = src.parent / f"{src.stem}{suffix}"
-            if srcfile.exists():
-                shutil.copy(str(srcfile), dst)
-    else:
-        for suffix in src_info.suffixes_extrafiles:
-            srcfile = src.parent / f"{src.stem}{suffix}"
-            dstfile = dst.parent / f"{dst.stem}{suffix}"
-            if srcfile.exists():
+
+    for suffix in src_info.suffixes_extrafiles:
+        srcfile = src.parent / f"{src.stem}{suffix}"
+        dstfile = (
+            dst / f"{src.stem}{suffix}"
+            if dst.is_dir()
+            else dst.parent / f"{dst.stem}{suffix}"
+        )
+        if srcfile.exists() and not dstfile.exists():
+            if keep_permissions:
                 shutil.copy(str(srcfile), dstfile)
+            else:
+                shutil.copyfile(srcfile, dstfile)
 
 
 def move(src: Union[str, "os.PathLike[Any]"], dst: Union[str, "os.PathLike[Any]"]):
